Carry out the planned renaming
