+++ core-bundle/src/Resources/contao/forms/FormCheckbox.php
@@ -11,14 +11,11 @@
 namespace Contao;
 
 /**
- * Class FormRadioButton
+ * Class FormCheckbox
  *
- * @property boolean $mandatory
- * @property array   $options
- *
- * @todo Rename to FormRadio in Contao 5.0
+ * @property array $options
  */
-class FormRadioButton extends Widget
+class FormCheckbox extends Widget
 {
 	/**
 	 * Submit user input
@@ -32,7 +29,7 @@ class FormRadioButton extends Widget
 	 *
 	 * @var string
 	 */
-	protected $strTemplate = 'form_radio';
+	protected $strTemplate = 'form_checkbox';
 
 	/**
 	 * Error message
@@ -46,30 +43,18 @@ class FormRadioButton extends Widget
 	 *
 	 * @var string
 	 */
-	protected $strPrefix = 'widget widget-radio';
+	protected $strPrefix = 'widget widget-checkbox';
 
 	/**
 	 * Add specific attributes
 	 *
-	 * @param string $strKey   The attribute key
+	 * @param string $strKey   The attribute name
 	 * @param mixed  $varValue The attribute value
 	 */
 	public function __set($strKey, $varValue)
 	{
 		switch ($strKey)
 		{
-			case 'mandatory':
-				if ($varValue)
-				{
-					$this->arrAttributes['required'] = 'required';
-				}
-				else
-				{
-					unset($this->arrAttributes['required']);
-				}
-				parent::__set($strKey, $varValue);
-				break;
-
 			case 'options':
 				$this->arrOptions = StringUtil::deserialize($varValue);
 				break;
@@ -91,7 +76,7 @@ class FormRadioButton extends Widget
 	/**
 	 * Return a parameter
 	 *
-	 * @param string $strKey The parameter name
+	 * @param string $strKey The parameter key
 	 *
 	 * @return mixed The parameter value
 	 */
@@ -106,18 +91,67 @@ class FormRadioButton extends Widget
 	}
 
 	/**
-	 * Check for a valid option (see #4383)
+	 * Check the options if the field is mandatory
 	 */
 	public function validate()
 	{
-		$varValue = $this->getPost($this->strName);
+		$mandatory = $this->mandatory;
+		$options = $this->getPost($this->strName);
 
-		if (!empty($varValue) && !$this->isValidOption($varValue))
+		// Check if there is at least one value
+		if ($mandatory && \is_array($options))
+		{
+			foreach ($options as $option)
+			{
+				if (\strlen($option))
+				{
+					$this->mandatory = false;
+					break;
+				}
+			}
+		}
+
+		$varInput = $this->validator($options);
+
+		// Check for a valid option (see #4383)
+		if (!empty($varInput) && !$this->isValidOption($varInput))
 		{
 			$this->addError($GLOBALS['TL_LANG']['ERR']['invalid']);
 		}
 
-		parent::validate();
+		// Add class "error"
+		if ($this->hasErrors())
+		{
+			$this->class = 'error';
+		}
+		else
+		{
+			$this->varValue = $varInput;
+		}
+
+		// Reset the property
+		if ($mandatory)
+		{
+			$this->mandatory = true;
+		}
+	}
+
+	/**
+	 * Return all attributes as string
+	 *
+	 * @param array $arrStrip An optional array with attributes to strip
+	 *
+	 * @return string The attributes string
+	 */
+	public function getAttributes($arrStrip=array())
+	{
+		// The "required" attribute only makes sense for single checkboxes
+		if ($this->mandatory && \count($this->arrOptions) == 1)
+		{
+			$this->arrAttributes['required'] = 'required';
+		}
+
+		return parent::getAttributes($arrStrip);
 	}
 
 	/**
@@ -158,7 +192,7 @@ class FormRadioButton extends Widget
 					array
 					(
 						'type'       => 'option',
-						'name'       => $this->strName,
+						'name'       => $this->strName . ((\count($this->arrOptions) > 1) ? '[]' : ''),
 						'id'         => $this->strId . '_' . $i,
 						'value'      => $arrOption['value'] ?? null,
 						'checked'    => $this->isChecked($arrOption),
@@ -206,8 +240,8 @@ class FormRadioButton extends Widget
 		foreach ($this->arrOptions as $i=>$arrOption)
 		{
 			$strOptions .= sprintf(
-				'<span><input type="radio" name="%s" id="opt_%s" class="radio" value="%s"%s%s%s <label id="lbl_%s" for="opt_%s">%s</label></span> ',
-				$this->strName,
+				'<span><input type="checkbox" name="%s" id="opt_%s" class="checkbox" value="%s"%s%s%s <label id="lbl_%s" for="opt_%s">%s</label></span> ',
+				$this->strName . ((\count($this->arrOptions) > 1) ? '[]' : ''),
 				$this->strId . '_' . $i,
 				$arrOption['value'] ?? null,
 				$this->isChecked($arrOption),
@@ -222,7 +256,7 @@ class FormRadioButton extends Widget
 		if ($this->strLabel)
 		{
 			return sprintf(
-				'<fieldset id="ctrl_%s" class="radio_container%s"><legend>%s%s%s</legend>%s<input type="hidden" name="%s" value=""%s%s</fieldset>',
+				'<fieldset id="ctrl_%s" class="checkbox_container%s"><legend>%s%s%s</legend>%s<input type="hidden" name="%s" value=""%s%s</fieldset>',
 				$this->strId,
 				($this->strClass ? ' ' . $this->strClass : ''),
 				($this->mandatory ? '<span class="invisible">' . $GLOBALS['TL_LANG']['MSC']['mandatory'] . ' </span>' : ''),
@@ -236,7 +270,7 @@ class FormRadioButton extends Widget
 		}
 
 		return sprintf(
-			'<fieldset id="ctrl_%s" class="radio_container%s">%s<input type="hidden" name="%s" value=""%s%s</fieldset>',
+			'<fieldset id="ctrl_%s" class="checkbox_container%s">%s<input type="hidden" name="%s" value=""%s%s</fieldset>',
 			$this->strId,
 			($this->strClass ? ' ' . $this->strClass : ''),
 			$this->strError,
