+++ core-bundle/src/Resources/contao/pages/PageRegular.php
@@ -234,7 +234,7 @@ class PageRegular extends Frontend
 		$this->Template->class = trim($objLayout->cssClass . ' ' . $objPage->cssClass);
 
 		// Execute AFTER the modules have been generated and create footer scripts first
-		$this->createFooterScripts($objLayout, $objPage);
+		$this->createFooterScripts($objPage, $objLayout);
 		$this->createHeaderScripts($objPage, $objLayout);
 	}
 
@@ -540,12 +540,10 @@ class PageRegular extends Frontend
 	/**
 	 * Create all footer scripts
 	 *
-	 * @param LayoutModel $objLayout
 	 * @param PageModel   $objPage
-	 *
-	 * @todo Change the method signature to ($objPage, $objLayout) in Contao 5.0
+	 * @param LayoutModel $objLayout
 	 */
-	protected function createFooterScripts($objLayout, $objPage = null)
+	protected function createFooterScripts($objPage, $objLayout)
 	{
 		$strScripts = '';
 		$nonce = ContaoFramework::getNonce();
