+++ core-bundle/tests/Twig/TwigIntegrationTest.php
@@ -17,7 +17,7 @@ use Contao\CoreBundle\Tests\TestCase;
 use Contao\CoreBundle\Twig\Extension\ContaoExtension;
 use Contao\CoreBundle\Twig\Inheritance\TemplateHierarchyInterface;
 use Contao\CoreBundle\Twig\Interop\ContextFactory;
-use Contao\FormTextField;
+use Contao\FormText;
 use Contao\System;
 use Contao\TemplateLoader;
 use Symfony\Component\Filesystem\Filesystem;
@@ -34,7 +34,7 @@ class TwigIntegrationTest extends TestCase
         (new Filesystem())->mkdir(Path::join($this->getTempDir(), 'templates'));
 
         $GLOBALS['TL_FFL'] = [
-            'text' => FormTextField::class,
+            'text' => FormText::class,
         ];
 
         $GLOBALS['TL_LANG']['MSC'] = [
@@ -61,10 +61,10 @@ class TwigIntegrationTest extends TestCase
         $content = "{{ strClass }}\n{{ strLabel }} {{ this.label }}\n {{ getErrorAsString }}";
 
         // Setup legacy framework and environment
-        (new Filesystem())->touch(Path::join($this->getTempDir(), 'templates/form_textfield.html5'));
-        TemplateLoader::addFile('form_textfield', 'templates');
+        (new Filesystem())->touch(Path::join($this->getTempDir(), 'templates/form_text.html5'));
+        TemplateLoader::addFile('form_text', 'templates');
 
-        $environment = new Environment(new ArrayLoader(['@Contao/form_textfield.html.twig' => $content]));
+        $environment = new Environment(new ArrayLoader(['@Contao/form_text.html.twig' => $content]));
         $environment->addExtension(new ContaoExtension($environment, $this->createMock(TemplateHierarchyInterface::class)));
 
         $container = $this->getContainerWithContaoConfiguration($this->getTempDir());
@@ -74,7 +74,7 @@ class TwigIntegrationTest extends TestCase
         System::setContainer($container);
 
         // Render widget
-        $textField = new FormTextField(['class' => 'my_class', 'label' => 'foo']);
+        $textField = new FormText(['class' => 'my_class', 'label' => 'foo']);
         $textField->addError('bar');
 
         $this->assertSame("my_class error\nfoo foo\n bar", $textField->parse());
